requirements.txt
requests
pytest-xdist
beautifulsoup4
lxml
